requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
python-dotenv
langgraph
langsmith
langchain-openai
llama-index-core
llama-index-embeddings-openai
openai
pandas
openpyxl
python-docx
python-pptx
pymupdf
pillow
matplotlib
wordcloud
numpy
camelot-py
opencv-python
paddleocr
paddlepaddle
pytesseract
pymupdf
openai
paddleocr
pytesseract
pillow
spacy
streamlit
dotenv

